requirements.txt
aiohttp==3.8.3
aiopg==1.3.5
aiosignal==1.3.1
aiosmtplib==1.1.7
alembic==1.8.1
amqp==5.1.1
anyio==3.6.2
asgi-lifespan==2.0.0
asgi-testclient==0.3.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
bidict==0.22.0
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
databases==0.6.1
datadog==0.44.0
defusedxml==0.7.1
dnspython==2.2.1
ecdsa==0.18.0
elasticsearch==7.17.7
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.85.1
fastapi-elasticsearch==0.5.3
fastapi-mail==1.2.0
filetype==1.1.0
flower==1.2.0
fonttools==4.38.0
fpdf2==2.5.7
frozenlist==1.3.3
gevent==22.10.1
greenlet==1.1.3.post0
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
humanize==4.4.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
netifaces==0.10.6
packaging==21.3
pika==1.3.1
Pillow==9.3.0
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-mock==3.10.0
python-dotenv==0.21.0
python-engineio==4.3.4
python-forge==18.6.0
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.7.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.39.0
starlette==0.20.4
svg.path==6.2
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
